Begin implementation of the event loop
In leiu of an asynchronous runtime, the current simulation engine
is based on an event loop incremented by a step function. At the
moment, it's only precise to the minute, but this isn't set in
stone. Events will be created both manually in the specification,
and automatically by the (currently unimplemented) simulation
profile.