simulation: Begin profile implementation
Profiles are used to specify an level of automatic events that
might occur during the simulation. At the moment, there are simply
low-, regular-, and high-activity profiles. This commit implements
the possible failures that might occur, but profiles also describe
the amount of power activity on the grid. This will be implemented
in a future commit.